Disable disk cache in safe mode

+++ lib/jekyll/cache.rb
@@ -17,14 +17,19 @@ module Jekyll
     def initialize(name)
       @@base_dir ||= File.expand_path(".jekyll-cache/Jekyll/Cache")
       @@caches ||= {}
+      @@safe ||= true
       @cache = @@caches[name] ||= {}
       @name = name
-      FileUtils.mkdir_p(path_to)
+      FileUtils.mkdir_p(path_to) if @@safe
     end
 
     def self.clear
       delete_cache_files
       @@caches.each_value(&:clear)
+    end
+
+    def self.safe(safe = true)
+      @@safe = safe unless defined?(@safe) && safe.nil?
     end
     # rubocop:enable Style/ClassVars
 
@@ -36,7 +41,7 @@ module Jekyll
     def [](key)
       return @cache[key] if @cache.key?(key)
       path = path_to(hash(key))
-      if File.file?(path) && File.readable?(path)
+      if @@safe && File.file?(path) && File.readable?(path)
         @cache[key] = load(path)
       else
         raise
@@ -46,29 +51,34 @@ module Jekyll
     def getset(key)
       return @cache[key] if @cache.key?(key)
       path = path_to(hash(key))
-      if File.file?(path) && File.readable?(path)
+      if @@safe && File.file?(path) && File.readable?(path)
         value = load(path)
       else
         value = yield
-        dump(path, value)
+        dump(path, value) if @@safe
       end
       @cache[key] = value
     end
 
     def []=(key, value)
       @cache[key] = value
-      path = path_to(hash(key))
-      dump(path, value)
+      if @@safe
+        path = path_to(hash(key))
+        dump(path, value)
+      end
     end
 
     def delete(key)
       @cache.delete(key)
-      path = path_to(hash(key))
-      File.delete(path)
+      if @@safe
+        path = path_to(hash(key))
+        File.delete(path)
+      end
     end
 
     def key?(key)
       return true if @cache.key?(key)
+      return false unless @@safe
       path = path_to(hash(key))
       File.file?(path) && File.readable?(path)
     end
@@ -99,11 +109,12 @@ module Jekyll
     end
 
     def delete_cache_files
-      FileUtils.rm_rf(path_to)
+      FileUtils.rm_rf(path_to) if @@safe
     end
 
     # rubocop:disable Security/MarshalLoad
     def load(path)
+      raise unless @@safe
       cached_file = File.open(path, "rb")
       value = Marshal.load(cached_file)
       cached_file.close
@@ -112,6 +123,7 @@ module Jekyll
     # rubocop:enable Security/MarshalLoad
 
     def dump(path, value)
+      return unless @@safe
       dir, _file = File.split(path)
       FileUtils.mkdir_p(dir)
       cached_file = File.open(path, "wb")
@@ -120,7 +132,7 @@ module Jekyll
     end
 
     def self.delete_cache_files
-      FileUtils.rm_rf(@@base_dir)
+      FileUtils.rm_rf(@@base_dir) if @@safe
     end
     private_class_method :delete_cache_files
   end
